[revisi] tinggal unit tests
