+++ services/user.go
@@ -1,14 +1,12 @@
 package services
 
 import (
-	"errors"
 	"pocket-message/dto"
 	"pocket-message/middleware"
 	"pocket-message/models"
 	"pocket-message/repositories"
 
 	"github.com/google/uuid"
-	"github.com/labstack/echo/v4"
 )
 
 func NewUserServices(db repositories.Database) UserServices {
@@ -16,50 +14,28 @@ func NewUserServices(db repositories.Database) UserServices {
 }
 
 type UserServices interface {
-	SignUp(echo.Context) error
-	Login(echo.Context) (dto.Login, error)
-	UpdateUsername(echo.Context) error
-	UpdatePassword(echo.Context) error
+	SignUp(user models.User) error
+	Login(user models.User) (dto.Login, error)
+	UpdateUsername(user models.User, token dto.Token) error
+	UpdatePassword(user models.User) error
 }
 
 type userServices struct {
 	repositories.Database
 }
 
-func (s *userServices) SignUp(c echo.Context) error {
-	var u models.User
-	err := c.Bind(&u)
-	if err != nil {
-		return err
-	}
-
-	if u.Username == "" {
-		return errors.New("username should not be empty")
-	}
-	if u.Password == "" {
-		return errors.New("password should not be empty")
-	}
+func (s *userServices) SignUp(u models.User) error {
 
 	u.UUID = uuid.New()
-	err = s.Database.SaveNewUser(u)
+
+	err := s.Database.SaveNewUser(u)
 	if err != nil {
 		return err
 	}
 
 	return nil
 }
-func (s *userServices) Login(c echo.Context) (dto.Login, error) {
-	var u models.User
-	err := c.Bind(&u)
-	if err != nil {
-		return dto.Login{}, err
-	}
-	if u.Username == "" {
-		return dto.Login{}, errors.New("username should not be empty")
-	}
-	if u.Password == "" {
-		return dto.Login{}, errors.New("password should not be empty")
-	}
+func (s *userServices) Login(u models.User) (dto.Login, error) {
 
 	user, err := s.Database.Login(u)
 	if err != nil {
@@ -77,43 +53,21 @@ func (s *userServices) Login(c echo.Context) (dto.Login, error) {
 
 	return result, nil
 }
-func (s *userServices) UpdateUsername(c echo.Context) error {
-	var u models.User
-	err := c.Bind(&u)
-	if err != nil {
-		return err
-	}
-	if u.Username == "" {
-		return errors.New("error, username should not be empty")
-	}
-	t, err := middleware.DecodeJWT(c)
-	if err != nil {
-		return err
-	}
+func (s *userServices) UpdateUsername(u models.User, t dto.Token) error {
+
 	u.UUID = t.UUID
 
-	err = s.Database.UpdateUsername(u)
+	err := s.Database.UpdateUsername(u)
 	if err != nil {
 		return err
 	}
 
 	return nil
 }
-func (s *userServices) UpdatePassword(c echo.Context) error {
-	var u models.User
-	err := c.Bind(&u)
-	if err != nil {
-		return err
-	}
 
-	if u.Username == "" {
-		return errors.New("username should not be empty")
-	}
-	if u.Password == "" {
-		return errors.New("password should not be empty")
-	}
+func (s *userServices) UpdatePassword(u models.User) error {
 
-	err = s.Database.UpdatePassword(u)
+	err := s.Database.UpdatePassword(u)
 	if err != nil {
 		return err
 	}
